rag-engine/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
httpx==0.28.1
chromadb==0.5.23
pypdf2==3.0.1
python-docx==1.1.2
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1
openai==1.58.1
aiofiles==24.1.0
sse-starlette==2.2.1
